+++ tests/symmetry/test_symmetry_pir.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from spgrep.irreps import enumerate_small_representations, enumerate_unitary_irreps
 from spgrep.rep.representation import is_representation
+from spgrep.symmetry.enumerate import enumerate_small_representations, enumerate_unitary_irreps
 from spgrep.symmetry.group import get_cayley_table, get_little_group
 from spgrep.utils import (
     NDArrayComplex,
