+++ tests/symmetry/test_symmetry_tensor.py
@@ -6,7 +6,7 @@ from spglib import get_symmetry_from_database
 
 from spgrep.rep.representation import is_representation
 from spgrep.symmetry.group import get_cayley_table
-from spgrep.tensor import apply_intrinsic_symmetry, get_symmetry_adapted_tensors
+from spgrep.symmetry.tensor import apply_intrinsic_symmetry, get_symmetry_adapted_tensors
 
 
 def get_standard_basis() -> list[np.ndarray]:
